Fixed a couple of compiler warnings for gcc (used in PyMELTS build)
* Other compiler warnings mostly due to NR
* Can ignore by adding flags to extra_compile_args in setup.py
  -Wno-array-bounds -Wno-unused-but-set-variable -Wno-strict-prototypes

+++ sources/read_write.c
@@ -1067,7 +1067,7 @@ int putOutputDataToFile(char *fileName)
       fprintf(output, "\n");
 #ifdef MAKE_TABLES
       if (tableSol[j] == NULL) {
-        size_t len = strlen(solids[j].label);
+        int len = (int) strlen(solids[j].label);
         char *nameOfFile = (char *) calloc((unsigned) (len+5), sizeof(char));
         strcpy(nameOfFile, solids[j].label);
         for (i=0; i<len; i++) if(nameOfFile[i] == ' ') nameOfFile[i] = '-';
@@ -1157,7 +1157,7 @@ int putOutputDataToFile(char *fileName)
       fprintf(output, "\n");
 #ifdef MAKE_TABLES
       if (tableSol[j] == NULL) {
-        size_t len = strlen(solids[j].label);
+        int len = (int) strlen(solids[j].label);
         char *nameOfFile = (char *) calloc((unsigned) (len+5), sizeof(char));
         strcpy(nameOfFile, solids[j].label);
         for (i=0; i<len; i++) if(nameOfFile[i] == ' ') nameOfFile[i] = '-';
